requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
cryptography==46.0.5
psycopg2-binary==2.9.11
pycparser==3.0
